add cooler control function

+++ settings.py
@@ -57,4 +57,4 @@ web_port = int(conf['WEB']['port'])
 time_interval_temp_humid_sec = int(conf['DATA_COLLECTION']['time_interval_temp_humid_sec'])
 
 time_interval_aircon_ai_sec = int(conf['AIRCON_AI']['time_interval_aircon_ai_sec'])
-target_temp = int(conf['AIRCON_AI']['target_temp'])
+target_temp = float(conf['AIRCON_AI']['target_temp'])
